requirements.txt
czifile
tifffile
numpy
opencv-python
Pillow
openpyxl  # Needed for Excel export